workaround mysterious warning
