--- src/shared/utils/types/route-data-response.ts
@@ -1,5 +1,0 @@
-import { RequestState } from "../../services/HttpService";
-
-export type RouteDataResponse<T extends Record<string, any>> = {
-  [K in keyof T]: RequestState<T[K]>;
-};
