--- src/shared/utils/types/route-data.ts
@@ -1,3 +1,0 @@
-import { RequestState } from "../../services/HttpService";
-
-export type RouteData = Record<string, RequestState<any>>;
