--- src/shared/utils/types/theme-color.ts
@@ -1,22 +1,0 @@
-export type ThemeColor =
-  | "primary"
-  | "secondary"
-  | "light"
-  | "dark"
-  | "success"
-  | "danger"
-  | "warning"
-  | "info"
-  | "blue"
-  | "indigo"
-  | "purple"
-  | "pink"
-  | "red"
-  | "orange"
-  | "yellow"
-  | "green"
-  | "teal"
-  | "cyan"
-  | "white"
-  | "gray"
-  | "gray-dark";
